// web/apps/webapp/src/queries/__generated__/locationMembersTab_paginatedLocationMembers_query.graphql.ts
export type locationMembersTab_paginatedLocationMembers_query$data = {
  readonly paginatedLocationMembers?: {
    readonly __id: string;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly id: string;
        readonly " $fragmentSpreads": FragmentRefs<"locationMemberCard_LocationMemberDetails">;
      };
    }>;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "locationMembersTab_paginatedLocationMembers_query";
};
export type locationMembersTab_paginatedLocationMembers_query$key = {
  readonly " $data"?: locationMembersTab_paginatedLocationMembers_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"locationMembersTab_paginatedLocationMembers_query">;
};